update modal handling scroll
